backend/requirements.txt
Django>=4.2
djangorestframework
gunicorn
python-dotenv
whitenoise
django-cors-headers

